+++ ninja-core-demo/src/main/java/controllers/ApplicationController.java
@@ -89,9 +89,9 @@ public class ApplicationController {
                              @Param("email") @Required String email) {
 
         if (validation.hasViolations()) {
-            return Results.json(validation.getFieldViolations("email"));
+            return Results.json().render(validation.getFieldViolations("email"));
         } else {
-            return Results.json(email);
+            return Results.json().render(email);
         }
     }
 
@@ -130,7 +130,7 @@ public class ApplicationController {
         Map<String, String> renderMap = Maps.newHashMap();
         renderMap.put("maliciousJavascript", maliciousJavascript);
 
-        return Results.html(renderMap);
+        return Results.html().render(renderMap);
 
     }