+++ ninja-core/src/site/markdown/developer/changelog.md
@@ -1,3 +1,10 @@
+Version X
+=========
+
+ * Removed deprecated and confusing methods in Results.java. Now we have only .json() .html() and so on.
+   https://github.com/reyez/ninja/issues/92#issuecomment-18015522
+   
+   
 Version 1.4.1
 =============